Made it so that the library itself (not most tests) can be compiled using MSVC. Thanks to Microsoft's preference to stick to only supporting C89 properly, complex numbers can't be used. Most changes in this commit are just including preprocessor switches to not compile parts of code which every other compiler can actually support when _MSC_BUILD is defined.

+++ source/cdouble/matrices/sparse_column_compressed.c
@@ -69,7 +69,7 @@ static jmtx_result ccs_insert_entry_at(jmtxz_matrix_ccs* mtx, uint32_t col, uint
         }
         mtx->indices = new_indices;
 
-        mtx->capacity = new_capacity;
+        mtx->capacity = (uint32_t)new_capacity;
     }
 
     //  Check for numer of values after the position
